Eliminate SlideUpPanel
